# revamp sean's bstream write optimizations