--- a/mdata/chunk/tsz/tsz_test.go
+++ b/mdata/chunk/tsz/tsz_test.go
@@ -2,6 +2,7 @@ package tsz
 
 import (
 	"math/rand"
+	"strconv"
 	"testing"
 )
 
@@ -24,68 +25,87 @@ func BenchmarkPushSeries4h(b *testing.B) {
 	b.Logf("Series4h size: %dB", len(s.Bytes()))
 }
 
-func benchmarkSeriesLong(b *testing.B, generator func(uint32) float64) {
-	s := NewSeriesLong(0)
-	N := uint32(b.N)
-	for i := uint32(1); i <= N; i++ {
-		s.Push(i, generator(i))
+var benchmarkSeriesLongChunkSizes = []int{1, 10, 30, 60, 120, 180, 240}
+
+func benchmarkSeriesLong(num int, generator func(uint32) float64) func(b *testing.B) {
+	return func(b *testing.B) {
+		var s *SeriesLong
+		for j := 0; j < b.N; j++ {
+			s = NewSeriesLong(0)
+			N := uint32(num)
+			for i := uint32(1); i <= N; i++ {
+				s.Push(i, generator(i))
+			}
+			s.Finish()
+		}
+		b.Logf("SeriesLong size: %d points in %dB, avg %.2f bytes/point, b.n %d", num, len(s.Bytes()), float64(len(s.Bytes()))/float64(num), b.N)
 	}
-	s.Finish()
-	b.Logf("SeriesLong size: %d points in %dB, avg %.2f bytes/point", N, len(s.Bytes()), float64(len(s.Bytes()))/float64(N))
 }
 
 func BenchmarkPushSeriesLongMonotonicIncreaseWithResets(b *testing.B) {
-	benchmarkSeriesLong(b, func(i uint32) float64 {
-		if i%10 == 0 {
-			return 0
-		} else if i%10 == 1 {
-			return 1
-		}
-		return float64(i) + 123.45
-	})
+	for _, num := range benchmarkSeriesLongChunkSizes {
+		b.Run(strconv.Itoa(num), benchmarkSeriesLong(num, func(i uint32) float64 {
+			if i%10 == 0 {
+				return 0
+			} else if i%10 == 1 {
+				return 1
+			}
+			return float64(i) + 123.45
+		}))
+	}
 }
 
 func BenchmarkPushSeriesLongMonotonicIncrease(b *testing.B) {
-	benchmarkSeriesLong(b, func(i uint32) float64 {
-		return float64(i) + 123.45
-	})
+	for _, num := range benchmarkSeriesLongChunkSizes {
+		b.Run(strconv.Itoa(num), benchmarkSeriesLong(num, func(i uint32) float64 {
+			return float64(i) + 123.45
+		}))
+	}
 }
 
 func BenchmarkPushSeriesLongSawtooth(b *testing.B) {
-	benchmarkSeriesLong(b, func(i uint32) float64 {
-		multiplier := 1.0
-		if i%2 == 0 {
-			multiplier = -1.0
-		}
-		return multiplier*123.45 + float64(i)/1000
-	})
+	for _, num := range benchmarkSeriesLongChunkSizes {
+		b.Run(strconv.Itoa(num), benchmarkSeriesLong(num, func(i uint32) float64 {
+			multiplier := 1.0
+			if i%2 == 0 {
+				multiplier = -1.0
+			}
+			return multiplier*123.45 + float64(i)/1000
+		}))
+	}
 }
 
 func BenchmarkPushSeriesLongSawtoothWithFlats(b *testing.B) {
-	benchmarkSeriesLong(b, func(i uint32) float64 {
-		multiplier := 1.0
-		if i%2 == 0 && i%100 != 0 {
-			multiplier = -1.0
-		}
-		return multiplier*123.45 + float64(i)/1000
-	})
+	for _, num := range benchmarkSeriesLongChunkSizes {
+		b.Run(strconv.Itoa(num), benchmarkSeriesLong(num, func(i uint32) float64 {
+			multiplier := 1.0
+			if i%2 == 0 && i%100 != 0 {
+				multiplier = -1.0
+			}
+			return multiplier*123.45 + float64(i)/1000
+		}))
+	}
 }
 
 func BenchmarkPushSeriesLongSteps(b *testing.B) {
-	benchmarkSeriesLong(b, func(i uint32) float64 {
-		multiplier := 1.0
-		if (i/100)%2 == 0 {
-			multiplier = -1.0
-		}
-		return multiplier*123.45 + float64(i)/1000
-	})
+	for _, num := range benchmarkSeriesLongChunkSizes {
+		b.Run(strconv.Itoa(num), benchmarkSeriesLong(num, func(i uint32) float64 {
+			multiplier := 1.0
+			if (i/100)%2 == 0 {
+				multiplier = -1.0
+			}
+			return multiplier*123.45 + float64(i)/1000
+		}))
+	}
 }
 
 func BenchmarkPushSeriesLongRealWorldCPU(b *testing.B) {
 	values := []float64{95.3, 95.7, 86.2, 95.0, 94.7, 95.4, 94.5, 94.0, 94.7, 95.0, 95.0, 93.8, 95.3, 95.4, 94.6, 83.8, 94.5, 94.5, 94.6, 92.0, 95.0, 89.6, 72.8, 72.1, 86.5, 94.9, 94.9, 93.9, 94.4, 95.4, 95.1, 93.7, 95.5, 95.4, 94.4, 93.2, 94.6, 95.5, 94.9, 94.1, 95.0, 95.5, 94.7, 93.7, 95.1, 96.6, 95.3, 94.0, 95.0, 95.2, 93.3, 94.2, 95.2, 94.9, 94.5, 95.3, 93.2, 95.4, 95.0, 95.2, 93.7}
-	benchmarkSeriesLong(b, func(i uint32) float64 {
-		return values[int(i)%len(values)]
-	})
+	for _, num := range benchmarkSeriesLongChunkSizes {
+		b.Run(strconv.Itoa(num), benchmarkSeriesLong(num, func(i uint32) float64 {
+			return values[int(i)%len(values)]
+		}))
+	}
 }
 
 func BenchmarkIterSeries4h(b *testing.B) {